+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This release reworks the scripting api and improves the scripting documentation.
 - Upd: some UI cleanups and improvements #909
 - Upd: Bootstrap 5.2.3
 - Upd: Documentation improvements
+- Upd: improve album cache creation performance #913
 
 ***